Reverted CF_SYSTEM_GROUP define, no longer needed
The root problem in ENT-5986 was that state dir lmdb files were
owned by the "wrong group" on solaris systems, "root" instead of "sys".
After some research it seems that masterfiles should be changed
to enforce a group of "root" instead as that is the default group
that the "root" user has for it's primary group.

Ticket: ENT-5986
Changelog: Title